COMP: Fix sign compararison warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=23091
git-svn-id: http://svn.slicer.org/Slicer4/trunk@23091 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Logic/vtkMRMLModelHierarchyLogic.cxx
@@ -254,7 +254,7 @@ void vtkMRMLModelHierarchyLogic::SetChildrenVisibility(vtkMRMLDisplayableHierarc
   // When there are many child nodes in a hierarchy then show/hide is much more efficient if batch processing is enabled.
   // However, if there are few nodes only then a full refresh at the end of a batch processing takes longer than doing
   // the update on each node separately.
-  bool batchProcess = (children.size() > vtkMRMLModelHierarchyLogic::ChildrenVisibilitySetBatchUpdateThreshold);
+  bool batchProcess = (children.size() > static_cast<unsigned int>(vtkMRMLModelHierarchyLogic::ChildrenVisibilitySetBatchUpdateThreshold));
   vtkMRMLScene* scene=NULL;
   if (batchProcess)
     {
